requirements.txt
absl-py                
art                    
astunparse             
blinker                
cachelib               
certifi                
cffi                   
charset-normalizer     
click                  
cmdstanpy              
colorama               
contourpy              
cycler                 
dnspython              
et_xmlfile             
filelock               
Flask                  
Flask-Cors             
flatbuffers            
fonttools              
future                 
gast                   
gevent                 
greenlet               
gunicorn               
h5py                   
holidays               
idna                   
importlib_resources    
itsdangerous           
Jinja2                 
joblib                 
kiwisolver             
Markdown               
markdown-it-py         
MarkupSafe             
matplotlib             
mdurl                  
ml_dtypes              
namex                  
numpy                  
opencv-python          
opencv-python-headless 
openpyxl               
packaging              
pandas                 
patsy                  
pdf2image              
pillow                 
pip                    
prophet                
protobuf               
pycparser              
Pygments               
pymongo                
PyMuPDF                
pyodbc                 
pyparsing              
PySocks                
pytesseract            
python-dateutil        
pytz                   
requests               
requests-file          
rich                   
scikit-learn           
scipy                  
setuptools             
six                    
stanio                 
statsmodels            
termcolor              
threadpoolctl          
tldextract             
tqdm                   
typing_extensions      
tzdata                 
urllib3                
waitress               
Werkzeug               
wfastcgi               
wheel                  
wrapt                  
xgboost                
zope.event             
zope.interface         